refactor: renamed auth pages to use naming convention
