+++ lib/pages/auth/auth_location.dart
@@ -1,8 +1,8 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:youniversity_app/layout/route_constants.dart';
-import 'package:youniversity_app/pages/auth/signin_page.dart';
-import 'package:youniversity_app/pages/auth/signup_page.dart';
+import 'package:youniversity_app/pages/auth/sign_in_page.dart';
+import 'package:youniversity_app/pages/auth/sign_up_page.dart';
 
 class AuthLocation extends BeamLocation<BeamState> {
   @override
